requirements.txt
Django>=3.0,<4.0
psycopg2-binary>=2.8
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
python-dotenv==0.19.1
faker==9.8.0
factory-boy==3.2.1
gunicorn==20.1.0